Consolidate dispatch & typecasting into processor
