+++ aspen/processor/typecasting.py
